Remove Data Layer Status Changes From Scheduler
This commit removes handling the status of the data layer from the
scheduler. A new go routine is now spun up to connect the data layer
and periodically (5 seconds) ping the service to ensure it is still
connected.

This completes the initial version of the health check API and the
monitoring of reliant services.